refactor: move upload to local_server
