+++ linkup-cli/src/commands/remote.rs
@@ -1,6 +1,6 @@
 use crate::{
     Result, services,
-    state::{ServiceTarget, State, upload_state},
+    state::{ServiceTarget, State},
 };
 
 use anyhow::anyhow;
@@ -63,7 +63,7 @@ pub async fn remote(args: &Args) -> Result<()> {
     }
 
     state.save()?;
-    upload_state(&state).await?;
+    services::local_server::upload_state(&state).await?;
 
     if args.all {
         println!("Linkup is routing all traffic to the remote servers");
